Gen AI Tutorial: Remove HF secret name
This change removes the Hugging Face secret name used during testing from the provided helm chart values files.

Because only the name of the secret (and not its contents) were present, this is not a data leak.

Additoinally, this change make all Hugging Face related variables beging w/ "HUGGING_FACE" and not "HF" or another value.